Fixes #3823: Reading Parquet from s3
