+++ extended/src/main/java/apoc/export/parquet/ParquetReadUtil.java
@@ -1,11 +1,9 @@
 package apoc.export.parquet;
 
-import apoc.ApocConfig;
+import apoc.util.CompressionAlgo;
+import apoc.util.FileUtils;
 import apoc.util.JsonUtil;
 import org.apache.commons.lang3.StringUtils;
-import org.apache.hadoop.conf.Configuration;
-import org.apache.hadoop.fs.Path;
-import org.apache.parquet.hadoop.util.HadoopInputFile;
 import org.apache.parquet.io.DelegatingSeekableInputStream;
 import org.apache.parquet.io.InputFile;
 import org.apache.parquet.io.SeekableInputStream;
@@ -33,7 +31,6 @@ import java.util.Map;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 
-import static apoc.util.FileUtils.changeFileUrlIfImportDirectoryConstrained;
 
 public class ParquetReadUtil {
 
@@ -168,19 +165,11 @@ public class ParquetReadUtil {
         };
     }
 
-
     public static InputFile getInputFile(Object source) throws IOException {
-        if (source instanceof String) {
-            ApocConfig.apocConfig().isImportFileEnabled();
-            String fileName = changeFileUrlIfImportDirectoryConstrained((String) source);
-            Path file = new Path(fileName);
-            return HadoopInputFile.fromPath(file, new Configuration());
-        }
-        return new ParquetStream((byte[]) source);
+        return new ParquetStream(FileUtils.inputStreamFor(source, null, null, CompressionAlgo.NONE.name()).readAllBytes());
     }
 
     public static ApocParquetReader getReader(Object source, ParquetConfig conf) {
-
         try {
             return new ApocParquetReader(getInputFile(source), conf);
         } catch (IOException e) {
